Fix conversion warnings

+++ src/instruction/bgez.cpp
@@ -10,7 +10,7 @@ Simulator::State Simulator::bgez(Instruction inst)
     auto rs = static_cast<int32_t>(m_state_iter->reg.at(op.rs));
 
     if (rs >= 0)
-        new_state.pc += static_cast<int32_t>(signExt(op.immediate, 16)) << 2;
+        new_state.pc += static_cast<int64_t>(signExt(op.immediate, 16)) << 2;
     else
         new_state.pc += 4;
